% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regF_gen.R
\name{regF_gen}
\alias{regF_gen}
\title{Creates the regression function for the mean}
\usage{
regF_gen(trend.dim, trend.pol, interaction, l, d)
}
\arguments{
\item{trend.dim}{which dimension should the trend follow: \code{"input"},
\code{"fidelity"}, or \code{"both"}.}

\item{trend.pol}{Which polynomial degree should the input mean trend have:
\code{"linear"} or \code{"quadratic"}.}

\item{interaction}{polynomial degree of the interaction between input trend
and fidelity trend of: \code{NULL}, \code{"linear"}, or \code{"quadratic"}.
\code{"linear"} or \code{"quadratic"}.}

\item{l}{convergence rate parameter, usually \code{l = 4}.}

\item{d}{input space dimension}
}
\value{
a function
}
\description{
Creates the regression function for the GP mean.
}
